requirements.txt
# Core testing framework
pytest>=7.4.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0
pytest-xdist>=3.3.1
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
pytest-rerunfailures>=12.0

# HTTP client and utilities
requests>=2.31.0
urllib3>=2.0.4
httpx>=0.24.1

# Data validation and serialization
jsonschema>=4.19.0
pydantic>=2.3.0
marshmallow>=3.20.1

# Configuration management
pyyaml>=6.0.1
python-dotenv>=1.0.0
configparser>=5.3.0

# Reporting and visualization
allure-pytest>=2.13.2
allure-python-commons>=2.13.2

# Mock and contract testing
wiremock>=2.6.0
pact-python>=2.0.0

# Data generation and utilities
faker>=19.6.2
factory-boy>=3.3.0
mimesis>=11.1.0

# Security and validation
cryptography>=41.0.4
jwt>=1.3.1
oauthlib>=3.2.2

# Performance testing
locust>=2.16.1

# Logging and monitoring
structlog>=23.1.0
colorlog>=6.7.0

# Utilities
click>=8.1.7
rich>=13.5.2
tabulate>=0.9.0
jinja2>=3.1.2
